Refactor keywords into their own classes, and ascal into its own class where multiple instances can exist in the same process without conflict

+++ AscalParameters.hpp
@@ -13,7 +13,7 @@
 
 class AscalParameters: public std::vector<Object*> {
 private:
-	int useCount;
+	uint16_t useCount;
 public:
 	AscalParameters();
 	AscalParameters& operator++();
